Moved Repository into database package
