+++ src/main/java/org/wahlzeit_revisited/api/filter/AuthenticationFilter.java
@@ -37,7 +37,7 @@ import jakarta.ws.rs.ext.Provider;
 import org.wahlzeit_revisited.api.auth.AccessRights;
 import org.wahlzeit_revisited.api.auth.PrincipalUser;
 import org.wahlzeit_revisited.api.dto.ErrorDto;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.database.repository.UserRepository;
 import org.wahlzeit_revisited.model.User;
 import org.wahlzeit_revisited.utils.SysLog;
