+++ src/main/java/org/wahlzeit_revisited/database/repository/UserRepository.java
@@ -19,7 +19,7 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-package org.wahlzeit_revisited.repository;
+package org.wahlzeit_revisited.database.repository;
 
 import jakarta.inject.Inject;
 import org.wahlzeit_revisited.model.User;
