+++ src/main/java/org/wahlzeit_revisited/model/Photo.java
@@ -20,7 +20,7 @@
 
 package org.wahlzeit_revisited.model;
 
-import org.wahlzeit_revisited.repository.Persistent;
+import org.wahlzeit_revisited.database.repository.Persistent;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
