+++ src/main/java/org/wahlzeit_revisited/model/User.java
@@ -21,7 +21,7 @@
 package org.wahlzeit_revisited.model;
 
 import org.wahlzeit_revisited.api.auth.AccessRights;
-import org.wahlzeit_revisited.repository.Persistent;
+import org.wahlzeit_revisited.database.repository.Persistent;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
