+++ src/main/java/org/wahlzeit_revisited/model/UserFactory.java
@@ -22,7 +22,7 @@
 package org.wahlzeit_revisited.model;
 
 import org.wahlzeit_revisited.api.auth.AccessRights;
-import org.wahlzeit_revisited.repository.PersistentFactory;
+import org.wahlzeit_revisited.database.repository.PersistentFactory;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
