+++ src/main/java/org/wahlzeit_revisited/service/PhotoFlagService.java
@@ -25,7 +25,7 @@ import jakarta.inject.Inject;
 import jakarta.ws.rs.ForbiddenException;
 import jakarta.ws.rs.NotFoundException;
 import org.wahlzeit_revisited.api.dto.PhotoDto;
-import org.wahlzeit_revisited.repository.PhotoRepository;
+import org.wahlzeit_revisited.database.repository.PhotoRepository;
 import org.wahlzeit_revisited.model.Photo;
 import org.wahlzeit_revisited.model.PhotoStatus;
 import org.wahlzeit_revisited.model.User;
