+++ src/test/java/org/wahlzeit_revisited/service/PhotoFlagServiceIT.java
@@ -10,8 +10,8 @@ import org.wahlzeit_revisited.model.PhotoFactory;
 import org.wahlzeit_revisited.model.PhotoStatus;
 import org.wahlzeit_revisited.model.User;
 import org.wahlzeit_revisited.model.UserFactory;
-import org.wahlzeit_revisited.repository.PhotoRepository;
-import org.wahlzeit_revisited.repository.UserRepository;
+import org.wahlzeit_revisited.database.repository.PhotoRepository;
+import org.wahlzeit_revisited.database.repository.UserRepository;
 
 import java.util.List;
 import java.util.Set;
